+++ lib/xue_hua_file_operations_method_channel.dart
@@ -186,4 +186,9 @@ class MethodChannelXueHuaFileOperations extends XueHuaFileOperationsPlatform {
       'identifier': ?identifier,
     });
   }
+
+  @override
+  Future<void> openAppSettings() async {
+    await _invoke<bool>('openAppSettings');
+  }
 }